zeroserver/pyproject.toml
[project]
name = "zeroserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pillow>=12.1.1",
    "pyzmq>=27.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
